requirements.txt
aenum==3.1.15
aiofiles==23.2.1
asgiref==3.7.2
asttokens==2.4.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bqplot==0.12.42
branca==0.7.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
channels==4.0.0
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
colour==0.1.5
comm==0.2.0
constantly==23.10.4
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
daphne==4.0.0
dataclasses-json==0.6.3
decorator==5.1.1
Django==5.0
django-vite==3.0.1
djangorestframework==3.14.0
earthengine-api==0.1.384
eerepr==0.0.4
executing==2.0.1
folium==0.15.1
fonttools==4.46.0
future==0.18.3
geemap==0.29.6
geocoder==1.38.1
google-api-core==2.15.0
google-api-python-client==2.111.0
google-auth==2.25.2
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-storage==2.14.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
httplib2==0.22.0
huepy==1.2.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
inotify==0.2.10
ipyevents==2.0.2
ipyfilechooser==0.6.0
ipyleaflet==0.18.1
ipython==8.18.1
ipytree==0.2.2
ipywidgets==8.1.1
jedi==0.19.1
Jinja2==3.1.2
jupyterlab-widgets==3.0.9
kiwisolver==1.4.5
livereload==2.6.3
markdown2==2.4.11
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.2
matplotlib-inline==0.1.6
msgpack==1.0.7
mypy-extensions==1.0.0
nose==1.3.7
numpy==1.26.2
oauthlib==3.2.2
packaging==23.2
pandas==2.1.4
parso==0.8.3
Pillow==10.1.0
plotly==5.18.0
prompt-toolkit==3.0.43
protobuf==4.25.1
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.17.2
pyOpenSSL==23.3.0
pyparsing==3.1.1
pyperclip==1.8.2
pyproj==3.6.1
pyshp==2.3.1
python-box==7.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
ratelim==0.1.6
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scooby==0.9.2
sentinelhub==3.10.0
service-identity==23.1.0
setuptools==69.0.2
shapely==2.0.2
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.3
tenacity==8.2.3
tifffile==2023.12.9
tomli==2.0.1
tomli_w==1.0.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.0
traittypes==0.2.1
Twisted==23.10.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
utm==0.7.0
vite==1.5.2
watchdog==3.0.0
wcwidth==0.2.12
whitenoise==6.6.0
widgetsnbextension==4.0.9
xyzservices==2023.10.1
zope.interface==6.1
